+++ src/migrations.ts
@@ -1,0 +1,67 @@
+import { PoolClient } from "pg";
+import { genPostgresAPI } from "./index";
+
+export interface Migration {
+  name: string;
+  exec: (client: PoolClient) => Promise<void>;
+}
+
+export interface MigrationConfig {
+  table?: string;
+}
+
+type MigrationTableCol = "id" | "created" | "name";
+
+interface MigrationTableRow {
+  id?:      number;
+  created?: Date;
+  name?:    string;
+}
+
+export const runPostgresMigrations = async (client: PoolClient, config: MigrationConfig, allMigrations: Migration[]): Promise<void> => {
+  const Migration = genPostgresAPI<MigrationTableCol, MigrationTableRow>(config.table || "migrations");
+  const initMigrationsTable = /*sql*/`CREATE TABLE IF NOT EXISTS "${config.table || "migrations"}" (
+    "id"      INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    "created" TIMESTAMP DEFAULT NOW(),
+    "name"    TEXT UNIQUE
+  );`;
+  try {
+    let migrations: MigrationTableRow[] = [];
+    try {
+      migrations = (await Migration.select(client, { cols: ["*"], })).rows;
+    } catch (error: any) {
+      if (error.code === "42P01") {
+        await client.query(initMigrationsTable, []);
+        await Migration.insert(client, {
+          cols: ["name"],
+          rows: [{ name: "init_db" }]
+        });
+        migrations = (await Migration.select(client, { cols: ["*"], })).rows;
+      } else {
+        throw error;
+      }
+    }
+    const activeMigrations: string[] = [];
+    for (const migration of migrations) {
+      activeMigrations.push(migration.name as string);
+    }
+    for (const migration of allMigrations) {
+      if (!activeMigrations.includes(migration.name)) {
+        await client.query("BEGIN");
+        await migration.exec(client);
+        await Migration.insert(client, {
+          cols: ["name"],
+          rows: [{ name: migration.name }]
+        });
+        await client.query("COMMIT");
+        activeMigrations.push(migration.name);
+        console.log(`Migration "${migration.name}" was executed successfully!`);
+      }
+    }
+  } catch (error) {
+    await client.query("ROLLBACK");
+    throw error;
+  } finally {
+    client.release();
+  }
+};