+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # SQMAP CHANGELOG
 
+> 0.2.0 - 2023.04.24 - Migrations
+* Added basic support for Postgres migrations
+* Properly exported the SQM utils
+
 > 0.1.0 - 2023.04.23 - Initial release
 * Postgres database support
 * Insert, select, update and delete queries support